vaisc-product-schema/archive-excel-converter/requirements.txt
# Excel Markdown Reformatter - Python Dependencies

# Core conversion engine
markitdown[all]>=0.1.3

# Configuration management
PyYAML>=6.0

# Excel file processing
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Data processing utilities
numpy>=1.20.0

# Optional: Enhanced CLI features
colorama>=0.4.0  # Colored terminal output
tqdm>=4.60.0     # Progress bars